Renamed Paper UI to org.eclipse.smarthome.ui.paper.
