+++ extensions/ui/org.eclipse.smarthome.ui.paper/.bowerrc
