+++ extensions/ui/org.eclipse.smarthome.ui.paper/.classpath
