+++ extensions/ui/org.eclipse.smarthome.ui.paper/.project
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
-	<name>org.eclipse.smarthome.ui.paperui</name>
+	<name>org.eclipse.smarthome.ui.paper</name>
 	<comment></comment>
 	<projects>
 	</projects>
