+++ extensions/ui/org.eclipse.smarthome.ui.paper/.settings/org.eclipse.jdt.core.prefs
