+++ extensions/ui/org.eclipse.smarthome.ui.paper/.settings/org.maven.ide.eclipse.prefs
