+++ extensions/ui/org.eclipse.smarthome.ui.paper/META-INF/MANIFEST.MF
@@ -1,6 +1,4 @@
 Manifest-Version: 1.0
-Private-Package: org.eclipse.smarthome.ui.paperui.internal
-Ignore-Package: org.eclipse.smarthome.ui.paperui.internal
 Bundle-Name: Eclipse SmartHome Paper UI
 Bundle-Vendor: Eclipse.org/SmartHome
 Bundle-Version: 0.8.0.qualifier
@@ -9,7 +7,7 @@ Bundle-License: http://www.eclipse.org/legal/epl-v10.html
 Import-Package: org.osgi.service.component,
  org.osgi.service.http,
  org.slf4j
-Bundle-SymbolicName: org.eclipse.smarthome.ui.paperui;singleton:=true
+Bundle-SymbolicName: org.eclipse.smarthome.ui.paper;singleton:=true
 Bundle-RequiredExecutionEnvironment: JavaSE-1.7
 Service-Component: OSGI-INF/*
 Bundle-ClassPath: patch/,.
