+++ extensions/ui/org.eclipse.smarthome.ui.paper/OSGI-INF/paperuiapp.xml
@@ -8,7 +8,7 @@
     http://www.eclipse.org/legal/epl-v10.html
 
 -->
-<scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.1.0" activate="activate" deactivate="deactivate" name="org.eclipse.smarthome.ui.internal.servlet.PaperUIApp">
-   <implementation class="org.eclipse.smarthome.ui.internal.servlet.PaperUIApp"/>
+<scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.1.0" activate="activate" deactivate="deactivate" name="org.eclipse.smarthome.ui.paper.internal.PaperUIApp">
+   <implementation class="org.eclipse.smarthome.ui.paper.internal.PaperUIApp"/>
    <reference bind="setHttpService" cardinality="1..1" interface="org.osgi.service.http.HttpService" name="HttpService" policy="static" unbind="unsetHttpService"/>
 </scr:component>
