+++ extensions/ui/org.eclipse.smarthome.ui.paper/about_files/ANGULARJS_LICENSE
