+++ extensions/ui/org.eclipse.smarthome.ui.paper/about_files/BOOTSTRAP_LICENSE
