+++ extensions/ui/org.eclipse.smarthome.ui.paper/about_files/JQUERY_LICENSE
