+++ extensions/ui/org.eclipse.smarthome.ui.paper/about_files/MASONRY_LICENSE
