+++ extensions/ui/org.eclipse.smarthome.ui.paper/about_files/SPRINTF_LICENSE
