+++ extensions/ui/org.eclipse.smarthome.ui.paper/about_files/TINYCOLOR_LICENSE
