+++ extensions/ui/org.eclipse.smarthome.ui.paper/bower.json
