+++ extensions/ui/org.eclipse.smarthome.ui.paper/build.properties
