+++ extensions/ui/org.eclipse.smarthome.ui.paper/package.json
