+++ extensions/ui/org.eclipse.smarthome.ui.paper/web/css/components.css
