+++ extensions/ui/org.eclipse.smarthome.ui.paper/web/js/controllers.setup.js
