+++ extensions/ui/org.eclipse.smarthome.ui.paper/web/js/extensions.js
