+++ extensions/ui/org.eclipse.smarthome.ui.paper/web/js/main.js
