+++ extensions/ui/org.eclipse.smarthome.ui.paper/web/js/services.repositories.js
