+++ extensions/ui/org.eclipse.smarthome.ui.paper/web/partials/configuration.html
