+++ extensions/ui/org.eclipse.smarthome.ui.paper/web/partials/control.html
