+++ extensions/ui/org.eclipse.smarthome.ui.paper/web/partials/dialog.bindinginfo.html
