+++ extensions/ui/org.eclipse.smarthome.ui.paper/web/partials/dialog.removething.html
