+++ extensions/ui/org.eclipse.smarthome.ui.paper/web/partials/dialog.scan.html
