+++ extensions/ui/org.eclipse.smarthome.ui.paper/web/partials/include.thingconfig.html
