+++ extensions/ui/org.eclipse.smarthome.ui.paper/web/partials/preferences.html
