+++ extensions/ui/org.eclipse.smarthome.ui.paper/web/partials/setup.html
