+++ features/org.eclipse.smarthome.feature.runtime.ui/feature.xml
@@ -65,5 +65,12 @@
          install-size="0"
          version="0.0.0"
          unpack="false"/>
+   
+   <plugin
+         id="org.eclipse.smarthome.ui.paper"
+         download-size="0"
+         install-size="0"
+         version="0.0.0"
+         unpack="false"/>
 
 </feature>
